+++ src/audio.rs
@@ -142,4 +142,8 @@ impl Recorder {
 
         Ok(buf.into_inner())
     }
+
+    pub fn sample_rate(&self) -> u32 {
+        self.sample_rate
+    }
 }